requirements.txt
Package                Version
---------------------- -------
alembic                1.14.0
bcrypt                 4.2.1
blinker                1.9.0
click                  8.1.8
Flask                  3.1.0
Flask-JWT-Extended     4.7.1
Flask-Migrate          4.0.7
Flask-SQLAlchemy       3.1.1
gunicorn               23.0.0
iniconfig              2.0.0
itsdangerous           2.2.0
Jinja2                 3.1.5
Mako                   1.3.8
MarkupSafe             3.0.2
marshmallow            3.23.2
marshmallow-sqlalchemy 1.1.0
packaging              24.2
pip                    24.3.1
pluggy                 1.5.0
psycopg2               2.9.10
PyJWT                  2.10.1
pytest                 8.3.4
python-dotenv          1.0.1
SQLAlchemy             2.0.36
typing_extensions      4.12.2
Werkzeug               3.1.3
